Split off wishlist expose action
